rename ws hook
